+++ src/components/Home.js
@@ -12,6 +12,7 @@ import {
   getDoc,
   updateDoc,
 } from '../firebase/methods.js';
+import { Post } from './Post.js';
 
 export const Home = () => {
   const user = getUser();
@@ -59,6 +60,8 @@ export const Home = () => {
   const buttonPublish = document.createElement('button');
   buttonPublish.textContent = 'Publicar';
   buttonPublish.className = 'buttonPublish button';
+  // contenedor de los post
+  const divContainerPost = document.createElement('div');
 
   formBoxProfile.appendChild(divUserName);
   formBoxProfile.appendChild(divPost);
@@ -68,57 +71,20 @@ export const Home = () => {
   divbuttonPublish.appendChild(buttonPublish);
   divContainer.appendChild(formBoxProfile);
   // Pizarra dinámica
-  const formPizarra = document.createElement('form');
-  formPizarra.className = 'profilePizarra';
-  const divPizarraName = document.createElement('div');
-  divPizarraName.className = 'pizarraName';
-  const spamUserPizarra = document.createElement('spam');
-  spamUserPizarra.className = 'spamPizarra';
-  const divPostPizarra = document.createElement('div');
-  divPostPizarra.className = 'divPostBoard';
-  const inputPizarra = document.createElement('input');
-  inputPizarra.className = 'postBoard';
-  inputPizarra.placeholder = 'aquí irá su comentario';
-  const divButtons = document.createElement('div');
-  divButtons.className = 'buttons';
-  const buttonLike = document.createElement('button');
-  buttonLike.textContent = '';
-  buttonLike.className = 'buttonLike buttonBoard';
-  const imgPlane = document.createElement('img');
-  imgPlane.src = '/img/planeLike.png';
-  imgPlane.className = 'imgPlane';
-  const buttonEdit = document.createElement('button');
-  buttonEdit.textContent = 'editar';
-  buttonEdit.className = 'buttonEdit buttonBoard';
-  const buttonDelete = document.createElement('button');
-  buttonDelete.textContent = 'delete';
-  buttonDelete.className = 'buttonDelete buttonBoard';
-
-  formPizarra.appendChild(divPizarraName);
-  formPizarra.appendChild(divPostPizarra);
-  formPizarra.appendChild(divButtons);
-  divPizarraName.appendChild(spamUserPizarra);
-  divPostPizarra.appendChild(inputPizarra);
-  divButtons.appendChild(buttonLike);
-  buttonLike.appendChild(imgPlane);
-  divButtons.appendChild(buttonEdit);
-  divButtons.appendChild(buttonDelete);
-  divContainer.appendChild(formPizarra);
+  
   // Lógica
   const btnPublic = divHome.querySelector('.buttonPublish');
   const inputValue = divHome.querySelector('.postHome');
   btnPublic.addEventListener('click', (e) => {
-    e.preventDefault(); console.log('input', inputValue.value);
-   
+    e.preventDefault();
+    console.log('input', inputValue.value);
     const docRef = addDoc(collection(dataBase, "post"), {
       post: inputValue.value,
     });
   });
+  const post = Post();
+  divContainerPost.appendChild(post);
+  divHome.appendChild(divContainerPost);
+
   return divHome;
 };
-
-/* addDoc(doc(dataBase, "c"), {
-      name: "Los Angeles",
-      state: inputValue.value,
-      country: "USA"
-    }) */
